restructured screens folders
